+++ src/BAT.jl
@@ -96,6 +96,9 @@ using IntervalSets: Domain
 import DomainSets
 using DomainSets: UnitInterval, UnitCube, Rectangle, FullSpace, RealNumbers
 
+import LazyReports
+using LazyReports: LazyReport, lazyreport, lazyreport!, lazytable
+
 using ChainRulesCore: AbstractTangent, Tangent, NoTangent, ZeroTangent, AbstractThunk, unthunk
 
 using Functors: fmap, @functor