+++ src/deprecations.jl
@@ -1,5 +1,9 @@
 # This file is a part of BAT.jl, licensed under the MIT License (MIT).
 
+
+@deprecate bat_report(obj...) lazyreport(obj...)
+
+
 @noinline function MaxDensityNelderMead(; kwargs...)
     Base.depwarn("`MaxDensityNelderMead(;kwargs...)` is deprecated, use `OptimAlg(;optalg = Optim.NelderMead, kwargs...)` instead.", :MaxDensityNelderMead)
     optalg = BAT.ext_default(pkgext(Val(:Optim)), Val(:NELDERMEAD_ALG))