+++ src/samplers/evaluated_measure.jl
@@ -114,13 +114,12 @@ _estimated_max_logd(measure::EvaluatedMeasure) = _estimated_max_logd(something(m
 # end
 
 
-function bat_report!(md::Markdown.MD, em::EvaluatedMeasure)
+function LazyReports.pushcontent!(rpt::LazyReport, em::EvaluatedMeasure)
     if !isnothing(em.samples)
-        bat_report!(md, em.samples)
+        lazyreport!(rpt, em.samples)
     end
     if !isnothing(em._generator)
-        bat_report!(md, em._generator)
+        lazyreport!(rpt, em._generator)
     end
-
-    return md
+    return nothing
 end